few adjustments to make it work with 1.11LTS   - moved __init.py__ into a views.py   - added MiddlewareMixin due to deprecation since 1.10
